fixed bug on some gcc versions

+++ includes/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: bchapman <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/11 19:25:21 by bchapman          #+#    #+#             */
-/*   Updated: 2019/04/08 11:47:36 by bchapman         ###   ########.fr       */
+/*   Updated: 2019/04/10 17:47:57 by bchapman         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned long	ft_strlen(const char *s)
+size_t ft_strlen(const char *s)
 {
-	unsigned long	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
